tests: Add tests for un-escaping spaces
Both remote and local path may contain spaces; in this case the spaces
are escaped as '\040'. Actual '\' in the path is escaped as '\134'.

We have code for un-escaping spaces, but no tests; Add the missing tests
based on output from /proc/mounts. Mark one of the tests are broken
since the current code can un-escape '\040' but fails for '\134'.

Change-Id: I772b7e7e48f888f9b9790cd6963d65cd86912699
Signed-off-by: Nir Soffer <nsoffer@redhat.com>