utils: Raise special error if termination fail
If terminating process failed because of unexpected error (for example
Popen.poll() raised), we use to log the exception and return silently.

This is a disastrous behavior for storage, when we want to ensure that a
process was terminated. If we could not terminate the process, a storage
job running the process must stay in running state to prevent engine
from starting the same operation on another machine.

Typically when we cannot handle an error we should not handle it. But
in this special case, letting the caller handle the original error (e.g.
OSError) is not useful since it does not have any context about this
error.

To allow the caller to handle this error, we raise TerminatingFailure
exception, including the process pid and the original error.

Storage code may like to panic or move the job into ABORTING state to
signal engine that an operation was not completed. Other code may just
log this error and continue.

Change-Id: I5cfa5434ce011b9185550884233b5b233026d13c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>