vmleases: Introduce the vmlease module
This module manage vm leases in the vmleases special volume.

When creating a vm lease, vdsm will allocate a lease slot in the
vmleases special volume. The mapping between vm id and lease offset is
managed by this module, and stored in the first lease slot of the
volume.

Adding and removing leases will be performed on the SPM (like adding and
removing a volume). When starting a VM, vdsm will lookup the lease
offset for the vm in vmleases volume.

Change-Id: I99b1fcff74240a568eee0b85c6f4beae84acbdb5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>