virt.vmdevices.core: Add Lease device
Previously virt.vmdevices.storage.Drive was creating lease device xml
for a drive. We want to add vm leases which are not attached to any
drive, so we need to separate the lease from the Drive class.

This patch adds new Lease device class and use it in Drive to generate
the drive lease element.

Unfinished: need to initialize vm lease devices in the vm device list
dict, and add the lease xml to the domain xml.

Change-Id: I10fc2be51eccdd93e4c2440008f22a4ba019c466
Signed-off-by: Nir Soffer <nsoffer@redhat.com>