tests: Create images in the same way as volumes
When vdsm creates a raw volume, it creates an empty sparse file in the
same size as the image, without running qemu-img create. When creating
qcow2 file, we run qemu-img create.

This patch change the way we create images for qcow2 images tests since
using qemu-img create with raw file can lead to unexpected results in
some tests environment (e.g. travis-ci) and we are interesting only in
the our use case.

Change-Id: I5e2e7c2897360ad777d746b75c78f422b3879bcc
Signed-off-by: Nir Soffer <nsoffer@redhat.com>