xleases: Prevent use of index during update
If the index metadata updating flag is set, loading an index from
storage must fail. This means that the index is in the middle of an
update, or an update operation was aborted. In this case the index must
be formatted or rebuilt from storage.

Users creating a LeasesVolume need to check for the new InvalidIndex
exception, matching either InvalidMetadata, or IndexIsUpdating
exceptions. Both exceptions has now a detailed error message explaining
how to recover from this situation.

Here is an example error when an index is updating:

    vdsm.storage.xlease.IndexIsUpdating: Index is updating or an update operation
    was aborted, the index must be formatted or rebuilt from storage:
    <IndexMetadata version=1, lockspace='lockspace', mtime=1481306351,
    updating=True at 0x7f4ee90a09b0>

Change-Id: I64105976e3e19abd960e83f412f781d65a726ddd
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>