xleases: Create and activate the external leases volume
Create the external leases volume when creating a new storage domain, or
upgrading storage domain from version 3 to version 4. Activate the new
volume in the flows activating the other special volumes.

This change introduce a new special volume, that must be created and
activated only on file and block storage domain v4. To support this we
use now helper functions to return the special volumes for the current
domain version and type.

StorageDomainManifest gained these APIs:

- supports_external_leases(version) - tell if the storage domain type
  can use external leases at that version.

- special_volumes(version) - the special volumes used by the storage
  domain.

StorageDomain gained these APIs:

- supports_external_leases(version) - redirect to the manifest class.

- external_leases_path() - return the path to the external lease volume

- create_external_leases() - create the special external leases volume
  or reuse existing volume. Called during upgrade to version 4.  This
  should be used also when creating a new storage domain, but this
  requires a risky refactoring.

- format_external_leases(lockspace, path) - format the external leases
  volume index, destroying existing information. Called when creating a
  new storage domain or when upgrading to storage domain version 4.

The new "xleases" lv was added to the vdsm lvm.rules udev rule, ensuring
correct ownership and permissions of the device.

In the upgrade flow we attempt to create and format the new volume. If
creating or formatting fail, we do not attempt to clean up anything
since removing a partly created volume will likely fail if storage
cannot be accessed. A partly created volume will be reused and
reformatted on the next upgrade attempt.

Currently we use xlease.DirectFile for both file and block storage.
This should be replaced later with storage type specific backend to
protect from blocking entire vdsm process if storage become
inaccessible.

Change-Id: If81093da1f6b553a9b1447561cbc6c83debf1a34
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>