virt: Support for memory hotunplug
With this patch we introduce support for DIMM based memory hotunplug.
Memory hotunplug is not guaranteed to succeed (or fail) after any amount of
time.  For that reason we introduce asynchronous device removal in this patch.
We react to the libvirt event rather than waiting for possible success
synchronously with some timeout.  Engine is supposed to detect contingent
disappearance of the memory device in its regular device polling.

Bug-Url: https://bugzilla.redhat.com/1228543
Change-Id: I9ee046adea2190de1667c8d43473d3792a5dde54
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>