volume: Operation without modifying legality
During some operations we are not interested in modifying volume
legality. For example, in cold merge and image upload, we set the
volume to illegal in the beginning of the flow, and set it to legal on
the end on the flow. The volume operation runs on another host when the
volume is illegal, and it should be left illegal after the volume
operation completed.

Add set_illegal flag, set to True by default. If the flag is set to
False, we don't modify volume legality during the operation. This make
it possible to use volume operations in cold merge.

Change-Id: I7795bfacff45f08ff8b228d63c31ddd7f11c37d9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>