xleases: Add DirectFile.size() method
For reading resources from existing external leases volume we need to
know the volume size. Add generic implementation that works for both
files and block devices.

We may need a smarter version later to detect a logical volume that
needs a refresh, or to optimize the file storage version to use
os.stat(), but for now this is good enough.

Fix the tests to create a more typical volume size, matching the way we
create the external leases volume for both block and file storage.

Change-Id: I17dbb9df2016037356feca1ddfbbf4db0abdf8e5
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>