bridge: Remove unused typefixups
Remove unused code for "fixing" some keys in vdsm response. I think this
was added in the past when Adam tried to create language bindings for
vdsm, and the motivation was to avoid reserved keywords like "type" and
"class", and replace them with safe words like "domainClass" and
"domainType". I don't think that these fixups belong to vdsm, they
should be part of the language binding.

Change-Id: Idbacccc1deb6922602306fd0d435198228a3ac83
Signed-off-by: Nir Soffer <nsoffer@redhat.com>