utils: Accept logger instead of logger name
Accepting a logger in utils.traceback is easier for callers, more
robust, and easier to test.  Now we can easily test that a thread is
logging a traceback if the thread function raises.

Note: since utils.traceback raises after logging the traceback to the
requested logger, the new test is logging unwanted traceback to
stderror. We can hide this traceback by monkeypatcing stderr during the
test but it will be handler in a better way in the next patch.

Change-Id: I78730a089569a848d365051486c1c488fcd91980
Signed-off-by: Nir Soffer <nsoffer@redhat.com>