pylint: Silence warning about __securityOverride
The @secured method adds **kwargs to all methods, and handles the
special __securityOverride argument for allowing secured methods when
an object is not secured. Unfortunately pylint is checking the methods
before they are wrapped, so it cannot detect the additional *kwargs.

Making this code easy to check statically requires a major rewrite; this
is a good idea anyway but not now. Silence the errors so we can check
this file and detect regressions.

Change-Id: Ibf665a8975e17b8cd369e7a1d28ae3cd8687e142
Signed-off-by: Nir Soffer <nsoffer@redhat.com>