pylint: Unify removeMetadata interface
In BlockVolumeManifest removeMetadata requires metaId argument. In
FileVolumeManifest it does not accept any argument, and in
VolumeManifest, it is not defined at all. This works in runtime since
this method is called only from the the concrete class with the correct
number of arguments, but this make pylint unhappy and makes the code
harder to understand.

Add the missing abstract method in VolumeManifest, and add an optional
metaId argument to FileVolumeManifest so the interface is similar.

Change-Id: Ie4b24bc5da5ca2dcf60151cf062afc5f356e6d5f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>