pylint: Remove unused metaoff instance variable
BlockVolumeManifest.metaoff was always None, so the value was never
used. BlockVolume had a read only property returning the unused value
from the menifest, and no code was acessing this property. The only
effect of this instance variable was generating pylint error, since it
is not defined in the VolumeManifest.

Change-Id: Ie8c305741887a2c6094a978126d4c06b03bbefb7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>