gluster: Fix error handling when loading libgfapi
We used ctypes.util.find_library() to locate the libgfapi file name. If
the library is not installed, this function returns None, but the return
value was not checked. Trying to locate later the glfs_new() function we
failed with a bogus error:

    AttributeError: /vdsm/.tox/imports/bin/python: undefined symbol: glfs_new

Change the call to use "libgfapi.so.0". Now if the library is missing we
fail with a clear message:

    OSError: libgfapi.so.0: cannot open shared object file: No such file or directory

Change-Id: If9fbf2a198d6934aea914b59afaa9a31a76ca7ec
Signed-off-by: Nir Soffer <nsoffer@redhat.com>