qemuimg: Replace CommandStream with Command
Now that operation.Command supports streaming, we can use it to process
qemu-img progress reports. This simplifies qemuimg.QemuImgOperation as
most of the logic is already implemented in operation.Command.

This also simplify and make caller code safer, as operation.Command()
cannot be used incorrectly, and does not need to be closed when done.

The qcow2_test module can run now on python 3, since we don't use
deathSignal any more. However turns out that qemuimg.ProgressCommand is
mixing bytes and strings, so this module is not compatible yet with
python 3. I marked the estimate tests as expected failure on python 3
instead of skipping them.

Change-Id: Icb29860e7ec8ebc260f129de2066515cc3229743
Signed-off-by: Nir Soffer <nsoffer@redhat.com>