tests: Remove global panic mocking
Global mocking of the panic function requires using a test runner. We
want to be able to run any tests using existing tests runners like
py.test. Tests that needs to mock panic.panic() should mock
monkeypatch it during the test.

The old panic tests was testing the panic mock, instead of the real
panic. Replace the test with a test testing the real code.

Change-Id: I8ceebdc2ae2d457e9be1b8121327f50a7fa4bfd0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>