concurrent: Log BaseException when threads finish
We logged an exception when thread fails with Exception, but
BaseExceptions are ignored silently. This may hide real errors raising
BaseException subclasses.

We log now a DEBUG message if a thread terminated with SystemExit or
KeybardInterrupt. These are unlikely, but are not user code error.
Anything else, including BaseException subclasses is considered as
critical failure and log a traceback.

Change-Id: I2570e083bb56a832fee978114ef6c7386e6860d6
Bug-Url: https://bugzilla.redhat.com/1469175
Signed-off-by: Nir Soffer <nsoffer@redhat.com>