asyncevent: Fix double complete calls on errors
BufferedReader is calling the complete callback when the reader is
closed in handle_close(). This handler is called both for normal
termination (the file descriptor was closed), or when an error occurred,
from handle_error().

Due to asyncore error handling mechanism, if the complete callback
raises, handle_error is called. This was leading to a second failure in
the complete callback, raising a new exception which is *not* handled by
asynccore, and is *not* handled by the EventLoop, terminating the
EventLoop.

Here is an example error, simulated by adding an assert in the
monitor.py:

2017-07-10 22:14:38,335+0300 ERROR (check/loop) [storage.asyncevent]
Unhandled error in <vdsm.storage.asyncevent.BufferedReader connected at
0x36715a8> (asyncevent:527)
Traceback (most recent call last):
  File "/usr/lib64/python2.7/asyncore.py", line 114, in readwrite
    obj.handle_close()
  File "/usr/lib/python2.7/site-packages/vdsm/storage/asyncevent.py", line 523, in handle_close
    self._complete(self._data)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/check.py", line 318, in _read_completed
    self._check_completed(rc)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/check.py", line 336, in _check_completed
    self._complete(result)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/monitor.py", line 495, in _pathChecked
    assert 0, "Simulating unexpected failure in _acquireHostId()"
AssertionError: Simulating unexpected failure in _acquireHostId()
2017-07-10 22:14:38,336+0300 ERROR (check/loop) [root] FINISH thread
    <Thread(check/loop, started daemon 139829069199104)> failed
    (concurrent:198)
Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/vdsm/common/concurrent.py", line 191, in run
    ret = func(*args, **kwargs)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/asyncevent.py", line 125, in run_forever
    self._run_once()
  File "/usr/lib/python2.7/site-packages/vdsm/storage/asyncevent.py", line 162, in _run_once
    asyncore.poll2(timeout, self._channels)
  File "/usr/lib64/python2.7/asyncore.py", line 201, in poll2
    readwrite(obj, flags)
  File "/usr/lib64/python2.7/asyncore.py", line 123, in readwrite
    obj.handle_error()
  File "/usr/lib/python2.7/site-packages/vdsm/storage/asyncevent.py", line 528, in handle_error
    self.handle_close()
  File "/usr/lib/python2.7/site-packages/vdsm/storage/asyncevent.py", line 523, in handle_close
    self._complete(self._data)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/check.py", line 311, in _read_completed
    rc = self._proc.poll()
AttributeError: 'NoneType' object has no attribute 'poll'

This patch fix the double complete calls in the BufferedReader. I think
this is the root cause of this bug.

A later patch will fix the EventLoop so it does not terminate on I/O
errors from asyncore.

Change-Id: Iceea86a76838cd3fb7162f2c80ddd249a3740a8f
Bug-Url: https://bugzilla.redhat.com/1469175
Signed-off-by: Nir Soffer <nsoffer@redhat.com>