monitor: Separate monitor and checker threads
The checker thread was modifying the monitor thread state and submitting
monitor change state events. This risk blocking of the checker thread,
caused race when releasing the host id, and is a bad idea in general.

This patch separates these threads completely using an event queue. When
the checker thread completes a check, it submit a PATH_CHECKED event to
the monitor event queue. When some other thread want to stop a montior,
it submit a MONITOR_STOPPED event to the monitor event queue. The
monitor will wake up and handle submitted events in the monitor thread.
If the monitor is busy, it will handle the events when it becomes idle.

We can remove now the warnings about blocking the checker thread since
the checker thread cannot be blocked. We can also remove the lock
protecting the monitor status, since the status is modified now only
from the monitor thread.

This change also removes a possible delay when acquiring a host id if a
checker completed after the monitor. Previously, the monitor would try
to acquire the host id in the next cycle, but now the monitor would wake
up immediately and acquire the host id.

This change makes testing the monitor little harder, since now when we
trigger the checker complete callback, we have to wait until the monitor
thread wakes up and process the event. The FakeEvent is using now a
queue instead of list, so we can wait until the monitor thread submit a
status change event.

Change-Id: I73077014be6ab627b2dae22071acabe7c01aa6ae
Signed-off-by: Nir Soffer <nsoffer@redhat.com>