asyncevent: Improve asyncore integration
Add _poll_channels() method, based on asyncore.poll2() from python 3.7,
and 2.7. This method wait for events on the event loop channels, and
then add a callback for each ready channel, to run at the end of
_run_once.

With this change, a faulty handle_error() raising exception will not
cause the event loop to skip the next I/O callbacks. The error will be
logged and then the next callbacks will be handled.

This change also fix the following python bugs that may never be fixed
in 2.7 or even in 3.x, since asyncore was deprecated since version 3.6.

- http://bugs.python.org/issue30994
- http://bugs.python.org/issue30985
- http://bugs.python.org/issue30931

Change-Id: Ie047f083624f58b6d7b55970a52492f77d69dcf1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>