qcow2: Support compressed qcow2 uploads
If a compressed qcow2 image is uploaded, copying the image to block
storage is likely to fail, because we used the source image size for
allocating the destination image.

This patch enables qcow2 source image in qcow2.estimate_size() and use
it to estimate the destination file size when copying qcow2 image to
another storage domain.

Change-Id: I4ece05285a293da2495bcbebf19cad97fb2f32d8
Bug-Url: https://bugzilla.redhat.com/1470435
Signed-off-by: Nir Soffer <nsoffer@redhat.com>