udev: Introduce the MultipathMonitor interface
To avoid loosing events when checking the initial state of the system,
we must orchestrate listening for events, checking the initial state, and
starting the udev observer. This is not possible with the current design
accepting a callback for handling events.

This patch changes the MultipathListener to accept an object implementing
the MultipathMonitor interface instead of a callable. The interface will
be extended later to support correct startup process.

Change-Id: Idfd595b9432f20b5562c93960e73da4b75e456e1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>