Revert "tests: always close the client"
This commit introduced a regression in the tests, causing each test_echo
permutation to fail and skip after 30 seconds. This adds 3 minutes (50%)
to the build time.

In test_echo, JsonRpcClient implements stop as alias to close, so
utils.running() should already close the client. I'm not sure why
utlls.closing() is not working as expected in this case.

In test_event, the client was indeed not closed, but using utils.closing
is not correct. I hope the someone from the infra team will submit a
correct fix later.

Before this change:

0:03:03.655 stomp_test.StompTests
00:03:33.693     test_echo(1024, False)         SKIP
00:04:03.734     test_echo(1024, True)          SKIP
00:04:33.763     test_echo(16384, False)        SKIP
00:05:03.805     test_echo(16384, True)         SKIP
00:05:33.831     test_echo(4096, False)         SKIP
00:05:56.387     test_echo(4096, True)          SKIP

After this change:

00:03:16.393 stomp_test.StompTests
00:03:16.424     test_echo(1024, False)         OK
00:03:16.506     test_echo(1024, True)          OK
00:03:16.538     test_echo(16384, False)        OK
00:03:16.661     test_echo(16384, True)         OK
00:03:16.693     test_echo(4096, False)         OK
00:03:16.774     test_echo(4096, True)          OK

This reverts commit ec3920873a59132195784048b857bcf7874c6d44.

Change-Id: Ib7f1179c472eac0590d21c83deaeb4aa733e493e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>